adjust to moved LaunchError definition
