+++ tests/pytest_tests/unit_tests/test_launch/test_environment/test_aws.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 
 import pytest
 from botocore.exceptions import ClientError
-from wandb.errors import LaunchError
+from wandb.sdk.launch.utils import LaunchError
 from wandb.sdk.launch.environment.aws_environment import AwsEnvironment
