+++ tests/pytest_tests/unit_tests/test_launch/test_environment/test_gcp.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 
 import pytest
 from google.auth.exceptions import DefaultCredentialsError, RefreshError
-from wandb.errors import LaunchError
+from wandb.sdk.launch.utils import LaunchError
 from wandb.sdk.launch.environment.gcp_environment import GcpEnvironment
