+++ tests/pytest_tests/unit_tests/test_launch/test_registry/test_gcp_artifact_registry.py
@@ -1,7 +1,7 @@
 from unittest.mock import MagicMock
 
 import pytest
-from wandb.errors import LaunchError
+from wandb.sdk.launch.utils import LaunchError
 from wandb.sdk.launch.registry.google_artifact_registry import GoogleArtifactRegistry
