+++ tests/pytest_tests/unit_tests_old/tests_launch/test_launch_gcp.py
@@ -4,7 +4,7 @@ import pytest
 import wandb
 import wandb.sdk.launch.launch as launch
 from google.cloud import aiplatform
-from wandb.errors import LaunchError
+from wandb.sdk.launch.utils import LaunchError
 from wandb.sdk.launch.runner.gcp_vertex import get_gcp_config, run_shell
 from wandb.sdk.launch.utils import LaunchError
